requirement.txt
fastapi
uvicorn[standard]
python-dotenv
google-api-python-client
google-auth
pydantic
pytz